improve qt_dims tests
